🚚 separate components to other fiels
